Handled taxes
Added a class Catalog in system_manager that represents the catalog.
Added a class VehicleCatalog that represents a vehicle in the catalog.
Added an enum State in order_manager that represents the state of an order.
Added a class OrderLoanPayment that inherits from Order and represent an order that is payed with a loan.
Added a class OrderDownPayment that inherits from Order and represent an order down payed.
Added an abstract  class Order in order_manager that represents an order.
Added a class DataBaseTaxHandler in database_manager wich has a static method to parse tax.csv and return the appropriate tax value.
Added a database directory and a file tax.csv in it to know the value of the tax, based on countryFrom and countryTo.
Added verifications in App to see if the code works properly.

Signed-off-by: Ytek24 <youssef.tektn@gmail.com>