Created the HomeScene
Added a class that represents the menu bar.
Changed the HomeScene class to display pictures of vehicles and enabled to swipe between vehicles.
Added a class that respresents the filter bar.
Added an attribute for the path of the picture in VehiculCatalog and added getters and setters associated to it.
Added getters and setters in the Catalog class.
Added some examples in the App class to see if it works properly.

Signed-off-by: Ytek24 <youssef.tektn@gmail.com>