Cleaned the whole code
Changed the import of java.awt.Color by javafx.scene.paint.Color.
Added a class DataBaseVehicleHandler that reads a file and create the appropriate vehicle base on what's written.
Cleaned the class app thanks to the use of DataBaseVehicleHandler.vehicleDescriptionFromDataBase(catalog).
Chaged s/releaseDate/releasedDate and same for getters and setters.
Added getters to VehicleCatalog to get the values easier.

Signed-off-by: Ytek24 <youssef.tektn@gmail.com>