Added Filters
Added two new attributs to vehicleCatalog : catalogTitle and catalogDescription.
Changed everywhere these changes this changes had an effect : DataBaseVehicleHandler, VehicleCatalog, vehicle.csv.
Added a class FilterAndResearch that handles the creation of filters and provides lists of all that exists and has a usefull filter method.
Added Filters at the FilterBar.
Changed the Filter Bar constructor.
Added usefull fonctions that handles arrow buttons state : disable, enable; and that handles button events.
Created a class FilterItems in utils to ease the filter creation.
Added a bottom display that displays the researches, its handled by the class ResultsFilterAndReasearchViewer which also handles exceptions
in case of bad casts.
Changed values of vehicles informations in the file vehicle.csv.
Changed the display of the descritpion in VehicleDescriptionViewer.
Created a class PostLabel that represents the label of a behicle after the filter action.
Changed VehicleDescriptionViewer to improve the visual aspect.

Signed-off-by: Ytek24 <youssef.tektn@gmail.com>