Added Filters
Added a class FilterAndResearch that handles the creation of filters and provides lists of all that exists.
Added Filters at the FilterBar.
Changed the Filter Bar constructor.
Added a bottom display that displays the researches.
Changed values of vehicles descriptions in the file vehicle.csv.
Changed the display of the descritpion in VehicleDescriptionViewer.

Signed-off-by: Ytek24 <youssef.tektn@gmail.com>