Handled customers
Created a file customer.csv that stores all information about customers.
Handle the creation of cusomers threw the read of customer.csv.
Created an abstract class Customer.
Created a class Individual that inherits from Customer.
Created a class Company that inherits from Customer.
Created a class CustomerAspect that has an instance of customer and manage the aspect of the customer.
Created a class DataBaseCustomerHandler that handles the read of customer.csv and creates an ArrayList of CustomerAspect.
Created a class CustomersUtil that handles usefull function for customer.
Changed HomeScene to handle the implementation of customers.
Changed almost all MenuBar to display the login and logout menu for customers.

Signed-off-by: Ytek24 <youssef.tektn@gmail.com>